+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to Mystic CMake will be documented here.
 
 ### Added
 
-- CMake Module: Moduke Importing Helper
-- CMake Module: Third Party Recursive Inclusion
+- CMake Module: Modules Recursive Inclusion Helper
+- CMake Module: Module Importing Helper
+- CMake Module: Third Party Recursive Inclusion Helper
 - CMake Module: Third Party Dependency Config Helper
 - CMake Module: Coverage Tool
 - CMake Module: ASCII Banner